backend/agentic-core/orchestration/requirements.txt
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
crewai-tools>=0.1.0
mcp>=0.1.0
python-dotenv>=1.0.0


